Refactored to fit into same category (projectiles)
