+++ src/main/java/com/example/demo/projectiles/UserProjectile.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.projectiles;
 
 public class UserProjectile extends Projectile {
